Add files via upload
The codes created represent two classes: "Product" and "Order". The "Product" class, which is a product
in the online store and contains a field such as id, name, price and availabilityInStock. The "Order" class
stores information about orders, including a set of products, their availability and the total availability of orders.
The "Order" class contains methods for adding products to orders, calculating results and detailed
applications of orders. In the main method, two products of the "Product" class and one order are combined.
Products are added to orders using the add Product methods. Finally, the
displayDetails method is called, which displays the order details on the screen.

(PL)